Modify models to allow setting of training data path
This is so that we can specify which training data to use - the full advertisment texts, or the ones that only contain the section that meets the recommendation.

Also add assessAccuracy script that runs all the assessments of model accuracy, so we can just run that when we're ready to get the final results.